+++ ui/src/components/collections/schema/SchemaFieldRelation.svelte
@@ -101,8 +101,10 @@
     {...$$restProps}
 >
     <svelte:fragment let:interactive>
+        <div class="separator" />
+
         <Field
-            class="form-field required {!interactive ? 'disabled' : ''}"
+            class="form-field required {!interactive ? 'readonly' : ''}"
             inlineError
             name="schema.{key}.options.collectionId"
             let:uniqueId
@@ -131,6 +133,8 @@
             </ObjectSelect>
         </Field>
 
+        <div class="separator" />
+
         <Field
             class="form-field form-field-single-multiple-select {!interactive ? 'disabled' : ''}"
             inlineError
@@ -143,6 +147,8 @@
                 bind:keyOfSelected={isSingle}
             />
         </Field>
+
+        <div class="separator" />
     </svelte:fragment>
 
     <svelte:fragment slot="options">