+++ ui/src/components/collections/schema/SchemaFieldSelect.svelte
@@ -52,6 +52,8 @@
     {...$$restProps}
 >
     <svelte:fragment let:interactive>
+        <div class="separator" />
+
         <Field
             class="form-field required {!interactive ? 'disabled' : ''}"
             inlineError
@@ -68,6 +70,9 @@
                 />
             </div>
         </Field>
+
+        <div class="separator" />
+
         <Field
             class="form-field form-field-single-multiple-select {!interactive ? 'disabled' : ''}"
             inlineError
@@ -80,6 +85,8 @@
                 bind:keyOfSelected={isSingle}
             />
         </Field>
+
+        <div class="separator" />
     </svelte:fragment>
 
     <svelte:fragment slot="options">